signal, counter, work_target's, work_source's to new headers
