+++ include/groute/worklist/work_kernels.cu.h
@@ -35,124 +35,14 @@
 #include <cub/grid/grid_barrier.cuh>
 
 #include <groute/worklist/work_queue.cu.h>
+#include <groute/worklist/work_source.cu.h>
+#include <groute/worklist/work_target.cu.h>
+#include <groute/device/signal.cu.h>
 
 #define GTID (blockIdx.x * blockDim.x + threadIdx.x)
 
 namespace groute
 {
-    namespace dev
-    {
-        template<typename TLocal, typename TRemote, typename DWCallbacks>
-        struct WorkTargetWorklist
-        {
-        private:
-            dev::Worklist<TLocal>& m_worklist;
-            DWCallbacks& m_callbacks;
-
-        public:
-            __device__ __forceinline__  WorkTargetWorklist(dev::Worklist<TLocal>& worklist, DWCallbacks& callbacks) : m_worklist(worklist), m_callbacks(callbacks) { }
-
-            __device__ __forceinline__ void append_work(const TLocal& work)
-            {
-                m_worklist.append_warp(work);
-            }
-
-            __device__ __forceinline__ void append_work(const TRemote& work)
-            {
-                m_worklist.append_warp(m_callbacks.unpack(work));
-            }
-        };
-
-        template<typename T, typename DWCallbacks>
-        struct WorkTargetWorklist < T, T, DWCallbacks >
-        {
-        private:
-            dev::Worklist<T>& m_worklist;
-
-        public:
-            __device__ __forceinline__  WorkTargetWorklist(dev::Worklist<T>& worklist, DWCallbacks& callbacks) : m_worklist(worklist) { }
-
-            __device__ __forceinline__ void append_work(const T& work)
-            {
-                m_worklist.append_warp(work);
-            }
-        };
-
-        template<typename TLocal, typename TRemote, typename DWCallbacks>
-        struct WorkTargetSplit
-        {
-        private:
-            dev::CircularWorklist<TLocal>& m_remote_input;
-            dev::CircularWorklist<TRemote>& m_remote_output;
-            DWCallbacks& m_callbacks;
-
-        public:
-            __device__ __forceinline__  WorkTargetSplit(dev::CircularWorklist<TLocal>& remote_input, dev::CircularWorklist<TRemote>& remote_output, DWCallbacks& callbacks) :
-                m_remote_input(remote_input), m_remote_output(remote_output), m_callbacks(callbacks) { }
-
-            __device__ __forceinline__ void append_work(const TLocal& unpacked)
-            {
-                SplitFlags flags = m_callbacks.on_send(unpacked);
-                if (flags & SF_Take)
-                {
-                    m_remote_input.prepend_warp(unpacked); // prepending to input 
-                }
-
-                if (flags & SF_Pass)
-                {
-                    // pack data
-                    TRemote packed = m_callbacks.pack(unpacked);
-                    m_remote_output.append_warp(packed); // appending  
-                }
-            }
-
-            __device__ __forceinline__ void append_work(const TRemote& packed)
-            {
-                // unpack data
-                TLocal unpacked = m_callbacks.unpack(packed);
-                SplitFlags flags = m_callbacks.on_send(unpacked);
-                if (flags & SF_Take)
-                {
-                    m_remote_input.prepend_warp(unpacked); // prepending to input 
-                }
-
-                if (flags & SF_Pass)
-                {
-                    m_remote_output.append_warp(packed); // appending  
-                }
-            }
-        };
-
-        template<typename T, typename DWCallbacks>
-        struct WorkTargetSplit < T, T, DWCallbacks >
-        {
-        private:
-            dev::CircularWorklist<T>& m_remote_input;
-            dev::CircularWorklist<T>& m_remote_output;
-            DWCallbacks& m_callbacks;
-
-        public:
-            __device__ __forceinline__  WorkTargetSplit(dev::CircularWorklist<T>& remote_input, dev::CircularWorklist<T>& remote_output, DWCallbacks& callbacks) :
-                m_remote_input(remote_input), m_remote_output(remote_output), m_callbacks(callbacks) { }
-
-            __device__ __forceinline__ void append_work(const T& unpacked)
-            {
-                SplitFlags flags = m_callbacks.on_send(unpacked);
-                if (flags & SF_Take)
-                {
-                    m_remote_input.prepend_warp(unpacked); // prepending to input 
-                }
-
-                if (flags & SF_Pass)
-                {
-                    // pack data
-                    T packed = m_callbacks.pack(unpacked);
-                    m_remote_output.append_warp(packed); // appending  
-                }
-            }
-        };
-    }
-
     // Three different kernel fusion working schemes: Never stop, Run N times, 
     // and Run for N cycles
